qtx_fishTest: seperation of concerns, move commands to a different source file
