+++ engine/NQ/fisheye/fishcmd.h
@@ -1,0 +1,4 @@
+#include <lua.h>
+#include "fisheye.h"
+
+void F_init_commands(struct _zoom *zoom_, struct _rubix *rubix_, lua_State *lua_);
