+++ engine/NQ/fisheye/fisheye.c
@@ -10,7 +10,6 @@
 #include "console.h"
 #include "cvar.h"
 #include "draw.h"
-#include "fisheye.h"
 #include "host.h"
 #include "mathlib.h"
 #include "quakedef.h"
@@ -19,9 +18,11 @@
 #include "sys.h"
 #include "view.h"
 
+#include "fisheye.h"
 #include "fishmem.h"
 #include "fishlens.h"
-#include "imageutil.h" //unwanted dependency
+#include "fishcmd.h"
+#include "imageutil.h"
 
 #include <lua.h>
 #include <lauxlib.h>
@@ -39,8 +40,6 @@
 // should be on.  It is used by other files for modifying behaviors that fisheye 
 // depends on (e.g. square refdef, disabling water warp, hooking renderer).
 qboolean fisheye_enabled;
-
-qboolean shortcutkeys_enabled;
 
 // This is a globally accessible variable that is used to set the fov of each
 // camera view that we render.
@@ -67,82 +66,9 @@ static struct _globe globe;
 
 static struct _lens lens;
 
-static struct _zoom {
+static struct _zoom zoom;
 
-   qboolean changed;
-
-   enum { ZOOM_NONE, ZOOM_FOV, ZOOM_VFOV, ZOOM_COVER, ZOOM_CONTAIN } type;
-
-   // desired FOV in degrees
-   int fov;
-
-   // maximum FOV width and height of the current lens in degrees
-   int max_vfov, max_fov;
-
-} zoom;
-
-static struct _rubix {
-
-   // boolean signaling if rubix should be drawn
-   qboolean enabled;
-
-   int numcells;
-   double cell_size;
-   double pad_size;
-
-   // We color a plate like the side of a rubix cube so we
-   // can see how the lens distorts the plates. (indicatrix)
-   //
-   // EXAMPLE:
-   //
-   //    numcells  = 3
-   //    cell_size = 2
-   //    pad_size  = 1
-   //
-   //  A globe plate is split into a grid of "units"
-   //  The squares that are colored below are called "cells".
-   //  You can see below that there are 3x3 colored cells,
-   //  since numcells=3.
-   //
-   //  You can see below that each cell is 2x2 units large,
-   //  since cell_size=2.
-   //
-   //  Finally, you can see the cells have 1 unit of padding,
-   //  since pad_size=1.
-   //
-   //    ---------------------------------------------------
-   //    |    |    |    |    |    |    |    |    |    |    |
-   //    |    |    |    |    |    |    |    |    |    |    |
-   //    |----|----|----|----|----|----|----|----|----|----|
-   //    |    |XXXXXXXXX|    |XXXXXXXXX|    |XXXXXXXXX|    |
-   //    |    |XXXXXXXXX|    |XXXXXXXXX|    |XXXXXXXXX|    |
-   //    |----|XXXXXXXXX|----|XXXXXXXXX|----|XXXXXXXXX|----|
-   //    |    |XXXXXXXXX|    |XXXXXXXXX|    |XXXXXXXXX|    |
-   //    |    |XXXXXXXXX|    |XXXXXXXXX|    |XXXXXXXXX|    |
-   //    |----|----|----|----|----|----|----|----|----|----|
-   //    |    |    |    |    |    |    |    |    |    |    |
-   //    |    |    |    |    |    |    |    |    |    |    |
-   //    |----|----|----|----|----|----|----|----|----|----|
-   //    |    |XXXXXXXXX|    |XXXXXXXXX|    |XXXXXXXXX|    |
-   //    |    |XXXXXXXXX|    |XXXXXXXXX|    |XXXXXXXXX|    |
-   //    |----|XXXXXXXXX|----|XXXXXXXXX|----|XXXXXXXXX|----|
-   //    |    |XXXXXXXXX|    |XXXXXXXXX|    |XXXXXXXXX|    |
-   //    |    |XXXXXXXXX|    |XXXXXXXXX|    |XXXXXXXXX|    |
-   //    |----|----|----|----|----|----|----|----|----|----|
-   //    |    |    |    |    |    |    |    |    |    |    |
-   //    |    |    |    |    |    |    |    |    |    |    |
-   //    |----|----|----|----|----|----|----|----|----|----|
-   //    |    |XXXXXXXXX|    |XXXXXXXXX|    |XXXXXXXXX|    |
-   //    |    |XXXXXXXXX|    |XXXXXXXXX|    |XXXXXXXXX|    |
-   //    |----|XXXXXXXXX|----|XXXXXXXXX|----|XXXXXXXXX|----|
-   //    |    |XXXXXXXXX|    |XXXXXXXXX|    |XXXXXXXXX|    |
-   //    |    |XXXXXXXXX|    |XXXXXXXXX|    |XXXXXXXXX|    |
-   //    |----|----|----|----|----|----|----|----|----|----|
-   //    |    |    |    |    |    |    |    |    |    |    |
-   //    |    |    |    |    |    |    |    |    |    |    |
-   //    ---------------------------------------------------
-
-} rubix;
+static struct _rubix rubix;
 
 // -------------------------------------------------------------------------------- 
 // |                                                                              |
@@ -159,25 +85,6 @@ void F_RenderView(void);
 // public introspection functions
 struct _globe* F_getGlobe(void);
 struct _lens* F_getLens(void);
-
-// console commands
-static void cmd_fisheye(void);
-static void cmd_help(void);
-static void cmd_lens(void);
-static void cmd_globe(void);
-static void cmd_fov(void);
-static void cmd_vfov(void);
-static void cmd_dumppal(void);
-static void cmd_rubix(void);
-static void cmd_rubixgrid(void);
-static void cmd_cover(void);
-static void cmd_contain(void);
-static void cmd_saveglobe(void);
-static void cmd_shortcutkeys(void);
-
-// console autocomplete helpers
-static struct stree_root * cmdarg_lens(const char *arg);
-static struct stree_root * cmdarg_globe(const char *arg);
 
 // lens builder timing functions
 static void start_lens_builder_clock(void);
@@ -200,8 +107,6 @@ static int LUAtoC_lens_forward(vec3_t ray, double *x, double *y);
 static int LUAtoC_globe_plate(vec3_t ray, int *plate);
 
 // functions to manage the data and functions in the Lua interpreter state
-static qboolean LUA_load_lens(void);
-static qboolean LUA_load_globe(void);
 static void LUA_clear_lens(void);
 static void LUA_clear_globe(void);
 
@@ -210,8 +115,6 @@ static qboolean lua_func_exists(const char* name);
 
 // zoom functions
 static qboolean calc_zoom(void);
-static void clear_zoom(void);
-static void print_zoom(void);
 
 // lens pixel setters
 static void set_lensmap_grid(int lx, int ly, int px, int py, int plate_index);
@@ -268,21 +171,7 @@ void F_Init(void)
 
    init_lua();
 
-   Cmd_AddCommand("fisheye", cmd_fisheye);
-   Cmd_AddCommand("f_help", cmd_help);
-   Cmd_AddCommand("f_dumppal", cmd_dumppal);
-   Cmd_AddCommand("f_rubix", cmd_rubix);
-   Cmd_AddCommand("f_rubixgrid", cmd_rubixgrid);
-   Cmd_AddCommand("f_cover", cmd_cover);
-   Cmd_AddCommand("f_contain", cmd_contain);
-   Cmd_AddCommand("f_fov", cmd_fov);
-   Cmd_AddCommand("f_vfov", cmd_vfov);
-   Cmd_AddCommand("f_lens", cmd_lens);
-   Cmd_SetCompletion("f_lens", cmdarg_lens);
-   Cmd_AddCommand("f_globe", cmd_globe);
-   Cmd_SetCompletion("f_globe", cmdarg_globe);
-   Cmd_AddCommand("f_saveglobe", cmd_saveglobe);
-   Cmd_AddCommand("f_shortcutkeys", cmd_shortcutkeys);
+   F_init_commands(&zoom, &rubix, lua);
 
    // defaults
    Cmd_ExecuteString("fisheye 1", src_command);
@@ -444,262 +333,6 @@ static void create_palmap(void){
 
 // -------------------------------------------------------------------------------- 
 // |                                                                              |
-// |                           CONSOLE COMMANDS                                   |
-// |                                                                              |
-// --------------------------------------------------------------------------------
-
-static void cmd_dumppal(void)
-{
-   dumppal(host_basepal);
-}
-
-static void cmd_rubix(void)
-{
-   rubix.enabled = !rubix.enabled;
-   Con_Printf("Rubix is %s\n", rubix.enabled ? "ON" : "OFF");
-}
-
-static void cmd_rubixgrid(void)
-{
-   if (Cmd_Argc() == 4) {
-      rubix.numcells = Q_atof(Cmd_Argv(1));
-      rubix.cell_size = Q_atof(Cmd_Argv(2));
-      rubix.pad_size = Q_atof(Cmd_Argv(3));
-      lens.changed = true; // need to recompute lens to update grid
-   }
-   else {
-      Con_Printf("RubixGrid <numcells> <cellsize> <padsize>\n");
-      Con_Printf("   numcells (default 10) = %d\n", rubix.numcells);
-      Con_Printf("   cellsize (default  4) = %f\n", rubix.cell_size);
-      Con_Printf("   padsize  (default  1) = %f\n", rubix.pad_size);
-   }
-}
-
-static void cmd_cover(void)
-{
-   clear_zoom();
-   zoom.type = ZOOM_COVER;
-}
-
-static void cmd_contain(void)
-{
-   clear_zoom();
-   zoom.type = ZOOM_CONTAIN;
-}
-
-static void cmd_fisheye(void)
-{
-   if (Cmd_Argc() < 2) {
-      Con_Printf("Currently: ");
-      Con_Printf("fisheye %d\n", fisheye_enabled);
-      Con_Printf("\nTry F_HELP for more options and commands.\n");
-      return;
-   }
-   fisheye_enabled = Q_atoi(Cmd_Argv(1)); // will return 0 if not valid
-   vid.recalc_refdef = true;
-}
-
-static void cmd_shortcutkeys(void)
-{
-   shortcutkeys_enabled = !shortcutkeys_enabled;
-   if (shortcutkeys_enabled) {
-      Con_Printf("Enabled Fisheye shortcut keys: 1-9 = Lenses, Y,U,I,O,P = Globes\n");
-      Cmd_ExecuteString("bind 1 \"f_lens panini\"", src_command);
-      Cmd_ExecuteString("bind 2 \"f_lens stereographic\"", src_command);
-      Cmd_ExecuteString("bind 3 \"f_lens hammer\"", src_command);
-      Cmd_ExecuteString("bind 4 \"f_lens winkeltripel\"", src_command);
-      Cmd_ExecuteString("bind 5 \"f_lens fisheye1\"", src_command);
-      Cmd_ExecuteString("bind 6 \"f_lens mercator\"", src_command);
-      Cmd_ExecuteString("bind 7 \"f_lens quincuncial\"", src_command);
-      Cmd_ExecuteString("bind 8 \"f_lens cube\"", src_command);
-      Cmd_ExecuteString("bind 9 \"f_lens debug\"", src_command);
-      Cmd_ExecuteString("bind y \"f_globe cube\"", src_command);
-      Cmd_ExecuteString("bind u \"f_globe cube_edge\"", src_command);
-      Cmd_ExecuteString("bind i \"f_globe trism\"", src_command);
-      Cmd_ExecuteString("bind o \"f_globe tetra\"", src_command);
-      Cmd_ExecuteString("bind p \"f_globe fast\"", src_command);
-   }
-   else {
-      Con_Printf("Disabled Fisheye shortcut keys\n");
-      Cmd_ExecuteString("bind 1 \"impulse 1\"", src_command);
-      Cmd_ExecuteString("bind 2 \"impulse 2\"", src_command);
-      Cmd_ExecuteString("bind 3 \"impulse 3\"", src_command);
-      Cmd_ExecuteString("bind 4 \"impulse 4\"", src_command);
-      Cmd_ExecuteString("bind 5 \"impulse 5\"", src_command);
-      Cmd_ExecuteString("bind 6 \"impulse 6\"", src_command);
-      Cmd_ExecuteString("bind 7 \"impulse 7\"", src_command);
-      Cmd_ExecuteString("bind 8 \"impulse 8\"", src_command);
-      Cmd_ExecuteString("unbind 9", src_command);
-      Cmd_ExecuteString("unbind y", src_command);
-      Cmd_ExecuteString("unbind u", src_command);
-      Cmd_ExecuteString("unbind i", src_command);
-      Cmd_ExecuteString("unbind o", src_command);
-      Cmd_ExecuteString("unbind p", src_command);
-   }
-}
-
-static void cmd_help(void)
-{
-   Con_Printf("-----------------------------\n");
-   Con_Printf("Welcome to the FISHEYE ADDON!\n");
-   Con_Printf("-> fisheye 1    (ENABLE)\n");
-   Con_Printf("-> fisheye 0    (DISABLE)\n");
-   Con_Printf("\n");
-   Con_Printf("-> f_lens <tab>    (CHANGE LENS)\n");
-   Con_Printf("-> f_fov <degrees> (SET FOV)\n");
-   Con_Printf("\n");
-   Con_Printf("-> f_<tab>         (MORE COMMANDS)\n");
-   Con_Printf("-----------------------------\n");
-}
-
-static void cmd_fov(void)
-{
-   if (Cmd_Argc() < 2) { // no fov given
-      Con_Printf("f_fov <degrees>: set horizontal FOV\n");
-      print_zoom();
-      return;
-   }
-
-   clear_zoom();
-
-   zoom.type = ZOOM_FOV;
-   zoom.fov = (int)Q_atof(Cmd_Argv(1)); // will return 0 if not valid
-}
-
-static void cmd_vfov(void)
-{
-   if (Cmd_Argc() < 2) { // no fov given
-      Con_Printf("f_vfov <degrees>: set vertical FOV\n");
-      print_zoom();
-      return;
-   }
-
-   clear_zoom();
-
-   zoom.type = ZOOM_VFOV;
-   zoom.fov = (int)Q_atof(Cmd_Argv(1)); // will return 0 if not valid
-}
-
-// lens command
-static void cmd_lens(void)
-{
-   if (Cmd_Argc() < 2) { // no lens name given
-      Con_Printf("f_lens <name>: use a new lens\n");
-      Con_Printf("Currently: %s\n", lens.name);
-      return;
-   }
-
-   // trigger change
-   lens.changed = true;
-
-   // get name
-   strcpy(lens.name, Cmd_Argv(1));
-
-   // Display name
-   Con_Printf("f_lens %s", lens.name);
-
-   // load lens
-   lens.valid = LUA_load_lens();
-   if (!lens.valid) {
-      strcpy(lens.name,"");
-      Con_Printf("not a valid lens\n");
-   }
-
-   // execute the lens' onload command string if given
-   // (this is to provide a user-friendly default view of the lens (e.g. "f_fov 180"))
-   lua_getglobal(lua, "onload");
-   if (lua_isstring(lua, -1))
-   {
-      const char* onload = lua_tostring(lua, -1);
-      Cmd_ExecuteString(onload, src_command);
-
-      // display onload
-      Con_Printf("; %s\n", onload);
-   }
-   else {
-      // fail silently for now, resulting from two cases:
-      // 1. onload is nil (undefined)
-      // 2. onload is not a string
-      Con_Printf("\n");
-   }
-   lua_pop(lua, 1); // pop "onload"
-}
-
-// autocompletion for lens names
-static struct stree_root * cmdarg_lens(const char *arg)
-{
-   struct stree_root *root;
-
-   root = Z_Malloc(sizeof(struct stree_root));
-   if (root) {
-      *root = STREE_ROOT;
-
-      STree_AllocInit();
-      COM_ScanDir(root, "../lua-scripts/lenses", arg, ".lua", true);
-   }
-   return root;
-}
-
-static void cmd_saveglobe(void)
-{
-   if (Cmd_Argc() < 2) { // no file name given
-      Con_Printf("f_saveglobe <name> [full flag=0]: screenshot the globe plates\n");
-      return;
-   }
-
-   strncpy(globe.save.name, Cmd_Argv(1), 32);
-
-   if (Cmd_Argc() >= 3) {
-      globe.save.with_margins = Q_atoi(Cmd_Argv(2));
-   }
-   else {
-      globe.save.with_margins = 0;
-   }
-   globe.save.should = true;
-}
-
-static void cmd_globe(void)
-{
-   if (Cmd_Argc() < 2) { // no globe name given
-      Con_Printf("f_globe <name>: use a new globe\n");
-      Con_Printf("Currently: %s\n", globe.name);
-      return;
-   }
-
-   // trigger change
-   globe.changed = true;
-
-   // get name
-   strcpy(globe.name, Cmd_Argv(1));
-
-   // display name
-   Con_Printf("f_globe %s\n", globe.name);
-
-   // load globe
-   globe.valid = LUA_load_globe();
-   if (!globe.valid) {
-      strcpy(globe.name,"");
-      Con_Printf("not a valid globe\n");
-   }
-}
-
-// autocompletion for globe names
-static struct stree_root * cmdarg_globe(const char *arg)
-{
-   struct stree_root *root;
-
-   root = Z_Malloc(sizeof(struct stree_root));
-   if (root) {
-      *root = STREE_ROOT;
-
-      STree_AllocInit();
-      COM_ScanDir(root, "../lua-scripts/globes", arg, ".lua", true);
-   }
-   return root;
-}
-
-// -------------------------------------------------------------------------------- 
-// |                                                                              |
 // |                    PURE COORDINATE CONVERTERS                                |
 // |                                                                              |
 // --------------------------------------------------------------------------------
@@ -780,26 +413,6 @@ static void init_lua(void)
 // |                           ZOOM FUNCTIONS                                     |
 // |                                                                              |
 // --------------------------------------------------------------------------------
-
-static void clear_zoom(void)
-{
-   zoom.type = ZOOM_NONE;
-   zoom.fov = 0;
-   zoom.changed = true; // trigger change
-}
-
-static void print_zoom(void)
-{
-   Con_Printf("Zoom currently: ");
-   switch (zoom.type) {
-      case ZOOM_FOV:     Con_Printf("f_fov %d", zoom.fov); break;
-      case ZOOM_VFOV:    Con_Printf("f_vfov %d", zoom.fov); break;
-      case ZOOM_COVER:   Con_Printf("f_cover"); break;
-      case ZOOM_CONTAIN: Con_Printf("f_contain"); break;
-      default:           Con_Printf("none");
-   }
-   Con_Printf("\n");
-}
 
 static qboolean calc_zoom(void)
 {
@@ -896,14 +509,13 @@ static qboolean calc_zoom(void)
    return true;
 }
 
-// TODO: excise any and all references to PCX; use external library for png instead
 // -------------------------------------------------------------------------------- 
 // |                                                                              |
 // |                           GLOBE SAVER FUNCTIONS                              |
 // |                                                                              |
 // --------------------------------------------------------------------------------
 
-// copied from WritePCXfile in NQ/screen.c
+// copied from WritePNGfile in NQ/screen.c
 // write a plate 
 static void WritePNGplate(char *filename, int plate_index, int with_margins){
    const byte *palette = host_basepal;
@@ -913,7 +525,7 @@ static void WritePNGplate(char *filename, int plate_index, int with_margins){
 static void save_globe(void)
 {
    int i;
-   char pcxname[32];
+   char pngname[32];
 
    globe.save.should = false;
 
@@ -921,10 +533,10 @@ static void save_globe(void)
 
    for (i=0; i<globe.numplates; ++i) 
    {
-      qsnprintf(pcxname, 32, "%s%d.png", globe.save.name, i);
-      WritePNGplate(pcxname, i, globe.save.with_margins);
+      qsnprintf(pngname, 32, "%s%d.png", globe.save.name, i);
+      WritePNGplate(pngname, i, globe.save.with_margins);
 
-    Con_Printf("Wrote %s\n", pcxname);
+    Con_Printf("Wrote %s\n", pngname);
    }
 
     D_DisableBackBufferAccess();	// for adapters that can't stay mapped in
@@ -1102,7 +714,7 @@ static int LUAtoC_globe_plate(vec3_t ray, int *plate)
 // |                                                                              |
 // --------------------------------------------------------------------------------
 
-static qboolean LUA_load_lens(void)
+qboolean LUA_load_lens(void)
 {
    // clear Lua variables
    LUA_clear_lens();
@@ -1195,7 +807,7 @@ static qboolean LUA_load_lens(void)
    return true;
 }
 
-static qboolean LUA_load_globe(void)
+qboolean LUA_load_globe(void)
 {
    // clear Lua variables
    LUA_clear_globe();
